refactor: Move dashboard page.tsx content to app directory
