+++ app/(dashboard)/dashboard/page.tsx
@@ -1,12 +1,12 @@
+import { cookies } from "next/headers";
+import Image from "next/image";
 import { redirect } from "next/navigation";
 
 import { authOptions } from "@/lib/auth";
 import { getCurrentUser } from "@/lib/session";
-import { Button } from "@/components/ui/button";
-import { DashboardHeader } from "@/components/dashboard/header";
-import { DashboardShell } from "@/components/dashboard/shell";
-import MailPage from "@/components/new-dashboard/page";
-import { EmptyPlaceholder } from "@/components/shared/empty-placeholder";
+import { isValidJSONString } from "@/lib/utils";
+import { Dashboard } from "@/components/new-dashboard/components/dashboard-1";
+import { accounts, mails } from "@/components/new-dashboard/data";
 
 export const metadata = {
   title: "Dasboard",
@@ -20,5 +20,45 @@ export default async function DashboardPage() {
     redirect(authOptions?.pages?.signIn || "/login");
   }
 
-  return <MailPage />;
+  const layout = cookies().get("react-resizable-panels:layout");
+  const collapsed = cookies().get("react-resizable-panels:collapsed");
+
+  const defaultLayout =
+    layout && isValidJSONString(layout.value)
+      ? JSON.parse(layout.value)
+      : undefined;
+  const defaultCollapsed =
+    collapsed && isValidJSONString(collapsed.value)
+      ? JSON.parse(collapsed.value)
+      : undefined;
+
+  return (
+    <>
+      <div className="md:hidden">
+        <Image
+          src="/examples/mail-dark.png"
+          width={1280}
+          height={727}
+          alt="Mail"
+          className="hidden dark:block"
+        />
+        <Image
+          src="/examples/mail-light.png"
+          width={1280}
+          height={727}
+          alt="Mail"
+          className="block dark:hidden"
+        />
+      </div>
+      <div className="hidden flex-col md:flex">
+        <Dashboard
+          accounts={accounts}
+          mails={mails}
+          defaultLayout={defaultLayout}
+          defaultCollapsed={defaultCollapsed}
+          navCollapsedSize={4}
+        />
+      </div>
+    </>
+  );
 }
