--- components/new-dashboard/page.tsx
@@ -1,51 +1,0 @@
-import { cookies } from "next/headers";
-import Image from "next/image";
-
-import { isValidJSONString } from "@/lib/utils";
-
-import { Dashboard } from "./components/dashboard-1";
-import { accounts, mails } from "./data";
-
-export default function MailPage() {
-  const layout = cookies().get("react-resizable-panels:layout");
-  const collapsed = cookies().get("react-resizable-panels:collapsed");
-
-  const defaultLayout =
-    layout && isValidJSONString(layout.value)
-      ? JSON.parse(layout.value)
-      : undefined;
-  const defaultCollapsed =
-    collapsed && isValidJSONString(collapsed.value)
-      ? JSON.parse(collapsed.value)
-      : undefined;
-
-  return (
-    <>
-      <div className="md:hidden">
-        <Image
-          src="/examples/mail-dark.png"
-          width={1280}
-          height={727}
-          alt="Mail"
-          className="hidden dark:block"
-        />
-        <Image
-          src="/examples/mail-light.png"
-          width={1280}
-          height={727}
-          alt="Mail"
-          className="block dark:hidden"
-        />
-      </div>
-      <div className="hidden flex-col md:flex">
-        <Dashboard
-          accounts={accounts}
-          mails={mails}
-          defaultLayout={defaultLayout}
-          defaultCollapsed={defaultCollapsed}
-          navCollapsedSize={4}
-        />
-      </div>
-    </>
-  );
-}
